+++ ChangeLog.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for OpenAPI plugins.
+- Support for extra sources.
 
 ### Changed
 
 - JSON Patch support is now a plugin.
+- Breaking: openapi directory is now `${source root}/openapi`
 
 
 ## [0.11.1] - 2024-08-15